+++ NEWS.md
@@ -1,6 +1,14 @@
+# timevis 0.1.2
+
+2016-07-26
+
+- fix bug: When re-rendering a timeline, the old data are not removed (#3)
+- removed the `getData` `getWindow` `getIds` `getSelected` parameters and instead just return that info always (#4)
+- add `fit` param to `timevis()` that determines if to fit the items on the timeline by default or not 
+
 # timevis 0.1.1
 
-2017-07-25
+2016-07-25
 
 - refactor and modularize Shiny app code
 - UI improvements to Shiny app
@@ -9,6 +17,6 @@
 
 # timevis 0.1
 
-2017-07-25
+2016-07-25
 
 Initial CRAN release